pyproject.toml
[project]
name = "propchain"
version = "1.0.0"
description = "Decentralized Real Estate Fractionalization Protocol on Algorand"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "py-algorand-sdk>=2.4.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pillow>=10.2.0",
    "pytesseract>=0.3.10",
    "spacy>=3.7.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "reportlab>=4.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
